+++ 1.20.1/gradle.properties
@@ -4,6 +4,5 @@ supported_minecraft_versions = >=1.20 <=1.20.1
 
 # These should be automatically updated, unless the environment
 #  variable "DISABLE_PROPERTIES_UPDATE" is set.
-loader_version               = 0.16.12
 yarn_version                 = 1.20.1+build.10
-fabric_api_version           = 0.92.5+1.20.1
+fapi_version                 = 0.92.5+1.20.1
