+++ 1.20.2/gradle.properties
@@ -4,6 +4,5 @@ supported_minecraft_versions = 1.20.2
 
 # These should be automatically updated, unless the environment
 #  variable "DISABLE_PROPERTIES_UPDATE" is set.
-loader_version               = 0.16.12
 yarn_version                 = 1.20.2+build.4
-fabric_api_version           = 0.91.6+1.20.2
+fapi_version                 = 0.91.6+1.20.2
