+++ 1.20.4/gradle.properties
@@ -4,6 +4,5 @@ supported_minecraft_versions = >=1.20.3 <=1.20.4
 
 # These should be automatically updated, unless the environment
 #  variable "DISABLE_PROPERTIES_UPDATE" is set.
-loader_version               = 0.16.12
 yarn_version                 = 1.20.4+build.3
-fabric_api_version           = 0.97.2+1.20.4
+fapi_version                 = 0.97.2+1.20.4
