+++ 1.21.1/gradle.properties
@@ -4,7 +4,5 @@ supported_minecraft_versions = >=1.21 <=1.21.1
 
 # These should be automatically updated, unless the environment
 #  variable "DISABLE_PROPERTIES_UPDATE" is set.
-loader_version               = 0.16.12
 yarn_version                 = 1.21.1+build.3
-## Only used for testmod
-fabric_api_version           = 0.115.4+1.21.1
+fapi_version                 = 0.115.4+1.21.1
