+++ 1.21.1/src/testmod/java/testmod/config/ModConfig.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 import static testmod.Testmod.*;
 
-public class ModConfig extends Config {
+public class ModConfig extends Config<ModConfig> {
 
     public String hello = "Hello!";
 
@@ -25,6 +25,11 @@ public class ModConfig extends Config {
     @Override
     protected String getName() {
         return MOD_ID;
+    }
+
+    @Override
+    public ModConfig getDefaultConfig() {
+        return null;
     }
 
     @Override
