+++ 1.21.5/gradle.properties
@@ -4,7 +4,5 @@ supported_minecraft_versions = >=1.21.2 <=1.21.5
 
 # These should be automatically updated, unless the environment
 #  variable "DISABLE_PROPERTIES_UPDATE" is set.
-loader_version               = 0.16.12
 yarn_version                 = 1.21.5+build.1
-## Only used for testmod
-fabric_api_version           = 0.119.9+1.21.5
+fapi_version                 = 0.119.9+1.21.5
