+++ 1.21.5/src/testmod/java/testmod/config/ModConfig.java
@@ -2,24 +2,18 @@ package testmod.config;
 
 import blue.endless.jankson.Comment;
 import top.offsetmonkey538.monkeylib538.config.Config;
-import top.offsetmonkey538.monkeylib538.config.Datafixer;
 
-import java.util.List;
+import java.util.HashMap;
+import java.util.Map;
 
 import static testmod.Testmod.*;
 
-public class ModConfig extends Config {
+public class ModConfig extends Config<ModConfig> {
 
     public String hello = "Hello!";
 
-    //@Comment("Now that's nice!")
-    //public int Number = 69;
-
-    //@Comment("this is the new number")
-    //public int coolNumber = 9;
-
-    @Comment("This is an even better number!")
-    public int wowSuchNumber = 42;
+    @Comment("Now that's nice!")
+    public int Number = 69;
 
 
     @Override
@@ -28,15 +22,7 @@ public class ModConfig extends Config {
     }
 
     @Override
-    protected int getConfigVersion() {
-        return 2;
-    }
-
-    @Override
-    protected List<Datafixer> getDatafixers() {
-        return List.of(
-                (original, jankson) -> original.put("coolNumber", original.get("Number")),
-                (original, jankson) -> original.put("wowSuchNumber", original.get("coolNumber"))
-        );
+    public ModConfig getDefaultConfig() {
+        return new ModConfig();
     }
 }
