+++ build.gradle
@@ -2,6 +2,7 @@ import net.fabricmc.loom.task.RemapJarTask
 import dex.plugins.outlet.v2.util.ReleaseType
 
 plugins {
+	id 'java'
 	id 'fabric-loom' version '1.10-SNAPSHOT' apply false
 	id 'io.github.dexman545.outlet' version '1.6.1' apply true
 	id 'com.modrinth.minotaur' version '2.+' apply false
@@ -9,7 +10,8 @@ plugins {
 
 outlet {
     maintainPropertiesFile = System.getenv("DISABLE_PROPERTIES_UPDATE") == null
-    propertiesData = [
+    mcVersionRange = "*"
+	propertiesData = [
             'loader_version': outlet.loaderVersion()
     ]
 }
@@ -24,7 +26,6 @@ allprojects {
 
 subprojects {
 	apply plugin: "fabric-loom"
-    apply plugin: "io.github.dexman545.outlet"
     apply plugin: "maven-publish"
 
     outlet {
@@ -44,7 +45,7 @@ subprojects {
 	dependencies {
 		minecraft "com.mojang:minecraft:${project.minecraft_version}"
 		mappings "net.fabricmc:yarn:${project.yarn_version}:v2"
-		modImplementation "net.fabricmc:fabric-loader:${project.loader_version}"
+		modImplementation "net.fabricmc:fabric-loader:${rootProject.loader_version}"
 
 		modImplementation "net.fabricmc.fabric-api:fabric-api:${project.fapi_version}"
 	}
@@ -125,7 +126,7 @@ configure(subprojects.findAll { it.name == "common" }) {
 	version = rootProject.mod_version
 	if (System.getenv("IS_DEBUG") == "true") version = "${version}+${rootProject.debugVersion}"
 
-	archivesBaseName = "common"
+	base.archivesName = "common"
 
 	configurations {
 		includeApi
@@ -145,7 +146,7 @@ configure(subprojects.findAll {it.name != "common"}) {
 	version = "${rootProject.mod_version}+${project.minecraft_version}"
 	if (System.getenv("IS_DEBUG") == "true") version = "${version}-${rootProject.debugVersion}"
 
-	archivesBaseName = "monkeylib538"
+	base.archivesName = "monkeylib538"
 
 	loom {
 		splitEnvironmentSourceSets()
