+++ common/gradle.properties
@@ -4,11 +4,9 @@ supported_minecraft_versions = *
 
 # These should be automatically updated, unless the environment
 #  variable "DISABLE_PROPERTIES_UPDATE" is set.
-loader_version               = 0.15.11
 yarn_version                 = 1.21.5+build.1
+fapi_version                 = 0.100.1+1.21
 
 # Dependencies
 # Jankson, check at https://github.com/falkreon/Jankson
 jankson_version              = 1.2.3
-
-fabric_api_version           = 0.100.1+1.21
