+++ gradle.properties
@@ -4,8 +4,10 @@ org.gradle.parallel          = true
 
 # These should be automatically updated, unless the environment
 #  variable "DISABLE_PROPERTIES_UPDATE" is set.
-loader_version               = 0.15.11
+loader_version               = 0.16.14
 
 # Mod Properties
 mod_version                  = 3.0.0+config-command-test
 maven_group                  = top.offsetmonkey538.monkeylib538
+yarn_version = 1.21.5+build.1
+fapi_version = 0.128.0+1.21.5
